Cleanup checksum() docs, change 64 format to commonly used base64.
Make the error message and the check use the same data.

Remove interfering newline from croak messages.

Make clear that the class of the object effects the checksum.

Remove references to details of the implementation.  The module
we're using under the hood has no bearing on how it works.

I don't think we need to explain what hex and base64 are.