Fix the "modification of read-only value" errors from non-threaded Perls.
Its like this.  For some reason lost to the mists of time, bless will not bless
a reference to a read only value.  So bless \"foo", $class isn't supposed to work.
This is just what our meta objects do.  For some unknown reason, this doesn't
trip when threads are turned on.  Possibly something to do with @_ magic.

The work around is to take a reference to the reference.  Its stupid and
probably slow.

A better thing to do might be to take advantage of field hashes to store
the meta object instead of recreating it every time.