Add basic subroutine and method signatures.
Its very basic, but it uses the emerging common syntax and behavior.  Its essentially
cut & coded from Method::Signatures::Simple with a slight modification to not grab
the invocant.

This little bit fixes a lot of potential syntax issues which would benefit from
using a callback function.  For example:

    $hash->each( def($key, $val) {
         print "$key => $val\n";
    });

    my @foo = @array->map( def($thing) { $thing++ } );

Of course, that requires getting anonymous declarations working.

Can't decide between 'def' and 'func'.  'func' is short and probably the clearest.  'def'
is short and has the weight of languages behind it (Ruby, Python, Groovy, Scala).  So
we use both and one will be deprecated.