fix recognition of true as a dependency
Build.PL creates its own MYMETA.yml because YAML::Tiny fails to parse the delivered one correctly. This change to the data affects how MYMETA.yml is created and makes sure it does so in a way that the resulting YAML file will properly recognize true as the dependency, and not 1.