Add $scalar->path to turn make a Path::Tiny object from $scalar.
The load time of Path::Tiny is up to 25% of perl5i, so we must load it
on demand.

I benchmarked the cost of "require Path::Tiny" every time and it
came out to be a significant difference for object construction.

The docs are very understated for what impact on the ease of
file manipulation this has.

Path::Tiny has some useful class methods.  cwd, rootdir, tempfile
and tempdir.  There's no way to access them through perl5i, you
have to use Path::Tiny directly.  This is less than ideal, but
ok for now.