Fix ->methods to work with constants (and perl 5.28)
Constants created by ‘use constant’ are store in the stash as scalar
references, whereas constants created by sub foo () { 42 } are stored
as subs-in-globs.

Before this commit, $class->mc->methods would return the latter kind
of constant, and skip the former, unless someone at some point hap-
pened to take a reference to one of the former before ->mc->methods
got called.

If we are going to be poking around in stashes like this, then we do
need to be are of what kind of stuff perl sneaks in there.  Constants
should not be treated inconsistently because of the internal rep-
resentation.

Also, if I get my way, most subs will start to be stored in stashes as
simple coderefs (which saves LOTS of memory), in perl 5.28.

This commit kills two stones with one bird: it handles constants cor-
rectly (an existing bug) and gets ready for the optimization that I am
almost certain will make it into perl 5.28.  We don’t need new tests
for the latter, because my optimization makes existing tests fail.

(For more info on the optimization, see
<https://rt.perl.org/Ticket/Display.html?id=132252#txn-1500037> and
also <https://rt.perl.org/Ticket/Display.html?id=129916>.)